# Various minor loop optimizations