The intention of these optimizations is to work towards further optimizing `div` (as per #52).

Results before:
```
test bench_decimal_add ... bench:      92,279 ns/iter (+/- 25,416)
test bench_decimal_div ... bench:  13,800,815 ns/iter (+/- 2,377,286)
test bench_decimal_mul ... bench:     105,153 ns/iter (+/- 26,713)
test bench_decimal_sub ... bench:     102,872 ns/iter (+/- 50,470)
```

Results after:
```
test bench_decimal_add ... bench:      85,074 ns/iter (+/- 43,969)
test bench_decimal_div ... bench:   8,591,217 ns/iter (+/- 939,923)
test bench_decimal_mul ... bench:      98,795 ns/iter (+/- 13,896)
test bench_decimal_sub ... bench:      95,592 ns/iter (+/- 12,895)
```

NB: Add/Mul/Sub are comparable, however Division is notably sub-10,000,000. 

There are still further optimizations to be made - the goal is to make it sub-1,000,000.