@paupino 2017-12-29
Note, these numbers come from a different machine than previous benchmarks. Will update once I know the updated numbers based on the consistent machine.