@paupino 2017-12-29
Updated numbers:
```
test bench_decimal_add ... bench:      72,839 ns/iter (+/- 1,796)
test bench_decimal_div ... bench:   3,349,243 ns/iter (+/- 131,503)
test bench_decimal_mul ... bench:      66,279 ns/iter (+/- 6,745)
test bench_decimal_sub ... bench:      91,092 ns/iter (+/- 16,187)
```